# sorbet-runtime: Speed up methods with no meaningful return type