--- a/gems/sorbet-runtime/lib/types/private/methods/call_validation.rb
+++ b/gems/sorbet-runtime/lib/types/private/methods/call_validation.rb
@@ -75,7 +75,7 @@ module T::Private::Methods::CallValidation
        method_sig.return_type.eql?(T::Types::SelfType::Private::INSTANCE) ||
        method_sig.return_type.is_a?(T::Types::TypeParameter) ||
        method_sig.return_type.is_a?(T::Types::TypeVariable) ||
-       (method_sig.return_type.is_a?(T::Types::Simple) && method_sig.return_type.raw_type == BasicObject))
+       (method_sig.return_type.is_a?(T::Types::Simple) && method_sig.return_type.raw_type.eql?(BasicObject)))
 
     returns_anything_method = all_args_are_simple && return_is_ignorable