… gems/sorbet-runtime/lib/types/private/methods/call_validation_2_7.rb