<!-- (optional) Explain your change, focusing on the details of the solution. This is a great place to call out user-visible changes. -->


### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

I'm trying to build a rubocop that prevents people from using
`.void.checked(:tests)`.

The suggestion is to use `.returns(T.anything).checked(:tests)` instead, but
that currently comes with a performance penalty, because there is a special case
in the runtime for methods that return `.void`.

In the process of speeding up `.returns(T.anything)` methods, I realized that we
can apply the same thing to all methods which Sorbet will skip runtime checking
for, including `T.untyped`, `T.attached_class`, `T.self_type`,
`T.type_parameter`, `BasicObject` and type members/templates.


### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

Before:

    sig {params(x: Example).void}: 139.688 ns
    sig {params(x: Example).returns(T.anything)}: 188.611 ns

    sig {params(x: T.nilable(Example)).void}: 180.063 ns
    sig {params(x: T.nilable(Example)).returns(T.anything)}: 200.054 ns

After:

    sig {params(x: Example).void}: 139.103 ns
    sig {params(x: Example).returns(T.anything)}: 139.717 ns

    sig {params(x: T.nilable(Example)).void}: 178.575 ns
    sig {params(x: T.nilable(Example)).returns(T.anything)}: 179.478 ns

Full output after:

    ❯ bex rake bench:typecheck
    Vanilla Ruby method call: 19.518 ns
    Vanilla Ruby is_a?: 22.928 ns
    T.unsafe: 20.456 ns
    T.must on non-nil: 27.755 ns
    T::Types::Simple#valid?: 35.288 ns
    T.nilable(Integer).valid?: 50.812 ns
    T.any(Integer, Float, T::Boolean).valid?: 178.889 ns
    T.let(..., Integer): 181.623 ns
    sig {params(x: Integer).void}: 143.413 ns
    sig {params(x: Integer, blk: T.proc.void)} -- block literal: 583.669 ns
    sig {params(x: Integer, blk: T.proc.void)} -- block pass: 504.053 ns
    sig {params(x: Integer, blk: T.nilable(T.proc.void))} -- block literal: 207.812 ns
    sig {params(x: Integer, blk: T.nilable(T.proc.void))} -- block pass: 156.218 ns
    T.let(..., T.nilable(Integer)): 633.972 ns
    sig {params(x: T.nilable(Integer)).void}: 177.717 ns
    T.let(..., Example): 169.639 ns
    sig {params(x: Example).void}: 139.103 ns
    sig {params(x: Example).returns(T.anything)}: 139.717 ns
    T.let(..., T.nilable(Example)): 636.255 ns
    sig {params(x: T.nilable(Example)).void}: 178.575 ns
    sig {params(x: T.nilable(Example)).returns(T.anything)}: 179.478 ns
    sig {params(s: Symbol, x: Integer, y: Integer).void} (with kwargs): 1.016 μs
    direct call Object#class: 20.054 ns
    .bind(example).call Object#class: 108.268 ns
    .bind_call(example) Object#class: 64.995 ns