@jez 2024-01-08
We have a policy of testing changes to Sorbet against Stripe's codebase before
merging them. I've kicked off a test run for the current PR.  When the build
finishes, I'll share with you whether or how it failed. Thanks!

Stripe employees can see the build results here:

→ https://go/builds/bui_PL4LDdV6bdZ2J9
→ https://go/builds/bui_PL4L1ZWGsY5DNV
→ https://go/builds/bui_PL4LvNONnpgyGj

This change also has sorbet-runtime changes. Those tests:

→ https://go/builds/bui_PL4S0slFE8qWQP