@jez 2024-01-08
We have a policy of testing changes to Sorbet against Stripe's codebase before
merging them. I've kicked off a test run for the current PR.  When the build
finishes, I'll share with you whether or how it failed. Thanks!

Stripe employees can see the build results here:

→ https://go/builds/bui_PL7D9xPYhxMQJc
→ https://go/builds/bui_PL7Dru9gAN57TG
→ https://go/builds/bui_PL7DIhpcwD9h83

This change also has sorbet-runtime changes. Those tests:

→ https://go/builds/bui_PL7KO7sFll1svw