@jez 2024-01-10
We have a policy of testing changes to Sorbet against Stripe's codebase before
merging them. I've kicked off a test run for the current PR.  When the build
finishes, I'll share with you whether or how it failed. Thanks!

Stripe employees can see the build results here:

→ https://go/builds/bui_PLrgP9MQYna38i
→ https://go/builds/bui_PLrgaTZm1oRzbT
→ https://go/builds/bui_PLrgvOX6iLm1lm

This change also has sorbet-runtime changes. Those tests:

→ https://go/builds/bui_PLrmqEPPIHHUpD