@jez 2024-01-10
We have a policy of testing changes to Sorbet against Stripe's codebase before
merging them. I've kicked off a test run for the current PR.  When the build
finishes, I'll share with you whether or how it failed. Thanks!

Stripe employees can see the build results here:

→ https://go/builds/bui_PLsPPcZQBiT4Ov
→ https://go/builds/bui_PLsPSOZKTBcIXc
→ https://go/builds/bui_PLsPG13wmIz9zO

This change also has sorbet-runtime changes. Those tests:

→ https://go/builds/bui_PLsUhOB1PjR3BZ